Refactoring parser class
