+++ src/main/java/seedu/duke/Duke.java
@@ -11,7 +11,14 @@ public class Duke {
      */
     public static void main(String[] args) {
         ui.printGreetingMessage();
-        Parser.programLogic();
+        boolean exitProgram = false;
+        while (!exitProgram) {
+            String input = ui.getUserMessage();
+            Parser.parseCommand(input);
+            if (Parser.getCommand(input).contains("bye")) {
+                exitProgram = true;
+            }
+        }
         ui.printByeMessage();
     }
 }
